% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{close_db_connection}
\alias{close_db_connection}
\title{Close Database Connection}
\usage{
close_db_connection(conn, use_dbi = TRUE)
}
\arguments{
\item{conn}{A database connection object}

\item{use_dbi}{Whether the connection was created with DBI (TRUE) or DatabaseConnector (FALSE)}
}
\description{
Closes a database connection.
}
